Review example after PR comments
remove setLogLevel
review doc
